// src/types/supabase.d.ts
export type Database = {
  public: {
    Tables: {
      favorite_plants: {
        Row: {
          id: string
          plant_id: number
          user_id: string
        }
        Insert: {
          id?: string
          plant_id: number
          user_id?: string
        }
        Update: {
          id?: string
          plant_id?: number
          user_id?: string
        }
        Relationships: [
          {
            foreignKeyName: "favorite_plants_plant_id_fkey"
            columns: ["plant_id"]
            isOneToOne: false
            referencedRelation: "plants"
            referencedColumns: ["id"]
          },
          {
            foreignKeyName: "favorite_plants_user_id_fkey"
            columns: ["user_id"]
            isOneToOne: false
            referencedRelation: "users"
            referencedColumns: ["id"]
          },
        ]
      }
      plant_care_conditions: {
        Row: {
          average_height_description: string
          average_height_max: number
          average_height_min: number
          average_height_unit: string
          fertilizer: string
          id: number
          placement: string
          season: string
          soil_description: string
          soil_ph_max: number
          soil_ph_min: number
          sunlight_description: string
          sunlight_hours_max: number
          sunlight_hours_min: number
          temperature_description: string
          temperature_max: number
          temperature_min: number
          temperature_unit: string
        }
        Insert: {
          average_height_description: string
          average_height_max: number
          average_height_min: number
          average_height_unit: string
          fertilizer: string
          id?: number
          placement: string
          season: string
          soil_description: string
          soil_ph_max: number
          soil_ph_min: number
          sunlight_description: string
          sunlight_hours_max: number
          sunlight_hours_min: number
          temperature_description: string
          temperature_max: number
          temperature_min: number
          temperature_unit: string
        }
        Update: {
          average_height_description?: string
          average_height_max?: number
          average_height_min?: number
          average_height_unit?: string
          fertilizer?: string
          id?: number
          placement?: string
          season?: string
          soil_description?: string
          soil_ph_max?: number
          soil_ph_min?: number
          sunlight_description?: string
          sunlight_hours_max?: number
          sunlight_hours_min?: number
          temperature_description?: string
          temperature_max?: number
          temperature_min?: number
          temperature_unit?: string
        }
        Relationships: [
          {
            foreignKeyName: "plant_care_conditions_id_fkey"
            columns: ["id"]
            isOneToOne: true
            referencedRelation: "plants"
            referencedColumns: ["id"]
          },
        ]
      }
// ...
    }
    Views: {
      [_ in never]: never
    }
    Functions: {
      [_ in never]: never
    }
    Enums: {
      [_ in never]: never
    }
    CompositeTypes: {
      [_ in never]: never
    }
  }
}

type PublicSchema = Database[Extract<keyof Database, "public">]

export type Tables<
  PublicTableNameOrOptions extends
    | keyof (PublicSchema["Tables"] & PublicSchema["Views"])
    | { schema: keyof Database },
  TableName extends PublicTableNameOrOptions extends { schema: keyof Database }
    ? keyof (Database[PublicTableNameOrOptions["schema"]]["Tables"] &
        Database[PublicTableNameOrOptions["schema"]]["Views"])
    : never = never,
> = PublicTableNameOrOptions extends { schema: keyof Database }
  ? (Database[PublicTableNameOrOptions["schema"]]["Tables"] &
      Database[PublicTableNameOrOptions["schema"]]["Views"])[TableName] extends {
      Row: infer R
    }
    ? R
    : never
  : PublicTableNameOrOptions extends keyof (PublicSchema["Tables"] &
        PublicSchema["Views"])
    ? (PublicSchema["Tables"] &
        PublicSchema["Views"])[PublicTableNameOrOptions] extends {
        Row: infer R
      }
      ? R
      : never
    : never